requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
et_xmlfile==2.0.0
fastapi==0.117.1
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-genai==2.0.10
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.29
Markdown==3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
motor==3.7.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
openai==1.108.1
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pymongo==4.15.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
sarvamai==0.1.20
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
twilio==9.8.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zstandard==0.25.0
